Move `TxSeq` into new transaction module hierarchy.
A recent refactoring moved all transaction-related types into the
following module hierarchy:

Cardano.Wallet.Primitive.Types.Tx

This change moves the `TxSeq` module (and related modules) into this
hierarchy.
